requirements.txt
flask==2.0.1
gunicorn==20.1.0
openai==1.3.0
python-dotenv==0.19.0
Werkzeug==2.0.1